refactor math utils
